refacotr : 필요 없는 필드 삭제 및 닉네임 필드 추가
로그인 세션이 가져야 할 정보 중 비밀번호 필드는 보안에 취약하고, 필요도 없으므로 삭제하고 사용자의 닉네임 필드를 추가함
